Updating pyth classes to use dataclass
This makes the code easier to read and makes the comparison less janky.

* Updating PythPriceInfo to use datclass and flake8 updates
* Updating PythPriceComponent to use a datclass
* Updating PythPriceAccount tests to use dataclass
* Update the binary bits to use base64 encoded bytes for brevity

Nice review suggestion from @tony-ricciardi. Not yet using dataclass
with PythProductAccount as there is some magic with the _prices attribute
and the properties without full coverage yet. The tests were janky and I'll
figure it out later when adding coverage for the various update bits.